fix: actualize readme, move helper package, fix small cli error
