+++ README.md
@@ -8,7 +8,7 @@
   </a>
 </p>
 
-<p align="center">Open-source Internal Developer Platform (IDP) for developers and non-specialists to manage multi-instance deployments, microservices and configurations in Kubernetes & Docker.
+<p align="center">dyrector.io is a self-hosted container management Platform.
 
 <br>
 <img src="docs/dyrectorio-readme-demo.gif" alt="dyrector.io demo GIF"/>
@@ -27,7 +27,7 @@
 
 ## Overview
 
-dyrector.io is an open-source deployment platform that helps software teams manage releases & deployments easily and efficiently. While non-specialists are enabled to manage these processes in a simplified, self-service manner, specialists can deploy and manage containerized apps faster through the platform.
+dyrector.io is an open-source container management platform that helps software teams to manage releases & deployments efficiently. While non-specialists are enabled to manage these processes in a simplified, self-service manner, specialists can customize and manage containerized apps faster through the platform.
 
 Our product is a platform for:
 
@@ -58,6 +58,7 @@ Join our Discord and connect with other members to share and learn together. If 
 -   Container Registry integrations
 -   Fine-grained RBAC (WIP)
 -   ChatOps & notification solutions
+-   Proud Ory Kratos users
 
 ## How it works
 
@@ -85,38 +86,41 @@ Both Docker and Kubernetes require specialized staff to manage. Via dyrector.io,
 
 Seamless testing whenever your team wants to test the application, without waiting for a SysAdmin to set up an environment.
 
-## Get Started
+## Getting Started
 
-dyrector.io CLI, is a command-line interface that lets you run and manage the whole dyrector.io project locally.
+Our CLI tool lets you run and manage the whole dyrector.io project's containers, you will have a fully-featured platform locally.
 
-This means you get a full-featured dyrector.io platform running locally on your machine.
+Stack:
+- UI Service (crux-ui)
+- Backend Service (crux)
+- PostgreSQL database
+- Authentication (Ory Kratos)
+- Migrations
+- SMTP mail test server
 
 ### Prerequirements
 
 -   Docker installed on your system (Podman works, too).
 -   Go Compiler to run the CLI from its source code. (Precompiled binaries are planned)
 
-### With CLI
+### Go install
 
-> **Note**
-> The CLI is only available from the codebase at the moment.
+1. Execute `go install github.com/dyrector-io/dyrectorio/golang/cmd/dyo@develop`
+2. Execute `dyo up`
+3. After you navigated to `localhost:8000` (this is the default traefik port) you will see a Login screen
+4. Register an account with whatever e-mail address you see fit (doesn't have to be valid one)
+5. Navigate to `localhost:4436` where you will find your mail as all outgoing e-mails will land here
+6. Open your e-mail message and using the link inside you can activate your account
+7. Happy deploying! 🎬
+
+### From Source
 
 1. Clone the repository to your local workdir with `git clone`
-2. Execute `go run ./golang/cmd/dyo up` in the project root
-3. Open the `localhost:8000` and use the platform
+2. Execute `make up` in the project root
+3. Open the `localhost:8000`
+4. Happy deploying! 🎬
 
-dyrector.io's command-line interface (CLI) lets you run a complete dyrector.io development environment locally with the following services: UI Service (crux-ui), Backend Service (crux), PostgreSQL databases, Authentication, Migrations, and SMTP mail server.
-
-#### Using the platform with CLI for demonstration or testing
-
-1. Execute `go run ./golang/cmd/dyo up` in the project root
-2. After you navigated to `localhost:8000` (this is the default traefik port) you will see a Login screen
-3. Register an account with whatever e-mail address you see fit (doesn't have to be valid one)
-4. Navigate to `localhost:4436` where you will find your mail as all outgoing e-mails will land here
-5. Open your e-mail message and using the link inside you can activate your account
-6. Enjoy!
-
-#### Using the platform with CLI for development
+## Development
 
 1. Read the CLI documentation first(see the end of this section)
 2. Decide which part of the project you want to work on, in this case it is crux, crux-ui or both
@@ -127,52 +131,11 @@ dyrector.io's command-line interface (CLI) lets you run a complete dyrector.io d
 7. Register an account with whatever e-mail address you see fit (doesn't have to be valid one)
 8. Navigate to `localhost:4436` where you will find your mail as all outgoing e-mails will land here
 9. Open your e-mail message and using the link inside you can activate your account
-10. Enjoy!
+10. Fruitful contributing! 🎬
 
 Read more about the CLI in the [documentation](https://docs.dyrector.io/get-started/cli).
 
-### With Docker compose files
-
-The easiest way to get started with the dyrector.io platform is by our premade [docker-compose](./web/docker-compose.dev.yaml) file.
-
-1. First you have to create `.env` from `.env.example` files in your local environment, in the following folders: `/web`, `/web/crux`, `/web/crux-ui`
-
-On _Windows_ and _Mac_ use `host.docker.internal` in the `.env` files for URLs.
-
-2. Build the dyrector.io platform dependencies, this step is only required once:
-
-```
-docker-compose -f web/docker-compose.dev.yaml build
-```
-
-3. To run the dyrector.io platform dependencies:
-
-```
-docker-compose -f web/docker-compose.dev.yaml up
-```
-
-If you are having problems starting `crux-kratos-migrate` change the line endings of `web/kratos/entrypoint.sh` from CRLF to LF. These problems include `ERROR: Encountered errors while bringing up the project.` while running `docker-compose up` or `': No such file or directory` in the container log.
-
-4. Now as all the infrastructure is ready, you need to install all the node.js dependencies in the specific projects' folder. To install all ( `web/crux/`, `web/crux-ui/` ) dependencies run `npm install`.
-
-_on Apple Silicon run `npm install --target_arch=x64`_
-
-5. In `crux/` folder you have to migrate the database and generate the Prisma client
-
-```
-npx prisma generate
-npx prisma migrate deploy
-```
-
-6. To start services run `npm start` in the followind folders `/web/crux` and `/web/crux-ui`
-
-7. Open the platform in `host.docker.internal:8000` and register
-
-8. In local our auth service is using a mock, so you can confirm your email in `http://localhost:4436/`
-
-9. Now you can log in and use the platform
-
-### Hosted version (SaaS)
+## Hosted version (SaaS)
 
 We are planning to support a hosted version in the near future.
