+++ golang/internal/mapper/grpc_test.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/docker/docker/api/types/container"
 	"github.com/dyrector-io/dyrectorio/golang/internal/config"
-	"github.com/dyrector-io/dyrectorio/golang/internal/helper/image"
 	"github.com/dyrector-io/dyrectorio/golang/internal/mapper"
 	builder "github.com/dyrector-io/dyrectorio/golang/pkg/builder/container"
+	"github.com/dyrector-io/dyrectorio/golang/pkg/helper/image"
 	"github.com/dyrector-io/dyrectorio/protobuf/go/agent"
 	"github.com/dyrector-io/dyrectorio/protobuf/go/common"
 	"github.com/stretchr/testify/assert"
