+++ golang/pkg/crane/k8s/deploy_facade.go
@@ -10,9 +10,9 @@ import (
 	v1 "github.com/dyrector-io/dyrectorio/golang/api/v1"
 	"github.com/dyrector-io/dyrectorio/golang/internal/dogger"
 	"github.com/dyrector-io/dyrectorio/golang/internal/grpc"
-	imageHelper "github.com/dyrector-io/dyrectorio/golang/internal/helper/image"
 	builder "github.com/dyrector-io/dyrectorio/golang/pkg/builder/container"
 	"github.com/dyrector-io/dyrectorio/golang/pkg/crane/config"
+	imageHelper "github.com/dyrector-io/dyrectorio/golang/pkg/helper/image"
 
 	"k8s.io/apimachinery/pkg/api/errors"
 )
