+++ golang/pkg/dagent/utils/dockerhelper.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/rs/zerolog/log"
 
-	dockerHelper "github.com/dyrector-io/dyrectorio/golang/internal/helper/docker"
 	"github.com/dyrector-io/dyrectorio/golang/internal/mapper"
+	dockerHelper "github.com/dyrector-io/dyrectorio/golang/pkg/helper/docker"
 	"github.com/dyrector-io/dyrectorio/protobuf/go/common"
 )
