+++ golang/pkg/helper/image/types.go
